modified:   src/main/java/control/ClienteController.java 	modified:   src/main/java/control/PedidosController.java 	modified:   src/main/java/control/ProdutoController.java 	modified:   src/main/java/control/VendasController.java 	modified:   src/main/java/dao/ClienteDAO.java 	modified:   src/main/java/dao/ItemDAO.java 	modified:   src/main/java/dao/PedidoDAO.java 	modified:   src/main/java/dao/ProdutoDAO.java 	modified:   src/main/java/model/Cliente.java 	modified:   src/main/java/model/Item.java 	modified:   src/main/java/model/Pedido.java 	modified:   src/main/java/model/Produto.java 	deleted:    src/main/java/model/ProdutoDAO.java 	modified:   target/classes/control/ClienteController.class 	modified:   target/classes/control/ItemController.class 	modified:   target/classes/control/PedidosController.class 	modified:   target/classes/control/ProdutoController.class 	modified:   target/classes/control/VendasController.class 	modified:   target/classes/dao/ClienteDAO.class 	modified:   target/classes/dao/ItemDAO.class 	modified:   target/classes/dao/PedidoDAO.class 	modified:   target/classes/dao/ProdutoDAO.class 	modified:   target/classes/model/Cliente.class 	modified:   target/classes/model/Item.class 	modified:   target/classes/model/Pedido.class 	modified:   target/classes/model/Produto.class 	deleted:    target/classes/model/ProdutoDAO.class

+++ src/main/java/control/ClienteController.java
@@ -1,6 +1,7 @@
 package control;
 
-import java.util.*;
+import java.util.List;
+import java.util.Scanner;
 
 import dao.ClienteDAO;
 import model.Cliente;
@@ -74,7 +75,7 @@ public class ClienteController {
         int op = 0;
         do {
             System.out.print("\nDigite o código do cliente \n(Zero p/sair): ");
-            long codigo = input.nextLong();
+            int codigo = input.nextInt();
             input.nextLine();
             if (codigo == 0) {
                 op = 0;
@@ -118,7 +119,7 @@ public class ClienteController {
     // opção 4
     private static void selectClientesById() {
         System.out.print("\nDigite o código do cliente: ");
-        Cliente cliente = ClienteDAO.selectClienteById(input.nextLong());
+        Cliente cliente = ClienteDAO.selectClienteById(input.nextInt());
         input.nextLine();
         if (cliente != null) {
             System.out.println(cliente);
